Add files via upload
This is the shiny app I made. It has the code to do some penalized regression, and some random forest. 

Based on the comments we got from the last task, I had an option for like using data pre- or post- 2009, since so many variables were like only recorded after 2009. So yeah. Also ignore most of the comments in this file bc they're from another project that I just copy and pasted a lot of this code from. Yeah. Vibes.